+++ data/signalisation-codification-rtp.json
@@ -1,0 +1,1 @@
+{"DESCRIPTION_RTP":{"0":"12- Balise","1":"18- Viaduc","2":"1- Tige et base","3":"10- F\u00fbt et feux","4":"8- Pilier","5":"7- Console","6":"13- Poteau en acier","7":"4- Poteau en bois","8":"11- Pot. bois\/feux","9":"20- F\u00fbt et potence","10":"2- Tige et manchon","11":"5- Cable a\u00e9rien","12":"16- Tige et base double","13":"19- D\u00e9lin\u00e9ateur","14":"21- Tige de borne fontaine","15":"3- F\u00fbt","16":"9- Autres","17":"6- Base amovible","18":"15- Manchon","19":"14- Poteau en barre m\u00e9tallique","20":"17- Tige et manchon double"}}
