+++ src/main/java/com/example/mtl_parking_finder/MtlParkingFinderApplication.java
@@ -1,0 +1,13 @@
+package com.example.mtl_parking_finder;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MtlParkingFinderApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MtlParkingFinderApplication.class, args);
+	}
+
+}
