+++ src/test/java/com/example/mtl_parking_finder/MtlParkingFinderApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.mtl_parking_finder;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MtlParkingFinderApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
